+++ changes.md
@@ -1,6 +1,20 @@
 **Api changes**
 
 <details>
+<summary>Added QueryParameter(s)</summary>
+
+- added query parameter `expand` to method `get /{projectKey}/in-store/key={storeKey}/cart-discounts`
+- added query parameter `sort` to method `get /{projectKey}/in-store/key={storeKey}/cart-discounts`
+- added query parameter `limit` to method `get /{projectKey}/in-store/key={storeKey}/cart-discounts`
+- added query parameter `offset` to method `get /{projectKey}/in-store/key={storeKey}/cart-discounts`
+- added query parameter `withTotal` to method `get /{projectKey}/in-store/key={storeKey}/cart-discounts`
+- added query parameter `where` to method `get /{projectKey}/in-store/key={storeKey}/cart-discounts`
+- added query parameter `/^var[.][a-zA-Z0-9]+$/` to method `get /{projectKey}/in-store/key={storeKey}/cart-discounts`
+- added query parameter `expand` to method `post /{projectKey}/in-store/key={storeKey}/cart-discounts`
+</details>
+
+
+<details>
 <summary>Added Enum(s)</summary>
 
 - added enum `approval-flow` to type `ReferenceTypeId`
@@ -9,6 +23,13 @@
 - added enum `approval-rule` to type `ChangeSubscriptionResourceTypeId`
 - added enum `approval-flow` to type `MessageSubscriptionResourceTypeId`
 - added enum `approval-rule` to type `MessageSubscriptionResourceTypeId`
+</details>
+
+
+<details>
+<summary>Changed MethodResponseBody(s)</summary>
+
+- :warning: changed response body for `200: application/json` of method `get /{projectKey}/in-store/key={storeKey}/cart-discounts` from type `CartDiscount` to `CartDiscountPagedQueryResponse`
 </details>